pyproject.toml
name = "edutainment-forge"
version = "0.1.0"
description = "Convert Microsoft Learn content into engaging podcast episodes"
authors = ["EdutainmentForge Team <team@edutainmentforge.dev>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/edutainmentforge/edutainmentforge"
repository = "https://github.com/edutainmentforge/edutainmentforge"
keywords = ["podcast", "education", "microsoft-learn", "tts", "audio"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"
python-dotenv = "^1.0.0"
azure-cognitiveservices-speech = {version = "^1.34.0", optional = true}
pyttsx3 = {version = "^2.90", optional = true}
pydub = {version = "^0.25.1", optional = true}
flask = {version = "^3.0.0", optional = true}

[tool.poetry.extras]
azure = ["azure-cognitiveservices-speech"]
local = ["pyttsx3"]
audio = ["pydub"]
web = ["flask"]
full = ["azure-cognitiveservices-speech", "pyttsx3", "pydub", "flask"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.0"
flake8 = "^6.0.0"
black = "^23.0.0"
isort = "^5.12.0"
pre-commit = "^3.4.0"
mypy = "^1.5.0"

[tool.poetry.scripts]
edutainment-forge = "src.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
